move file to gem folder. remove some extra lines. rename the module base class to DataMigrations
