+++ lib/generators/data_migration_install_generator.rb
@@ -2,8 +2,7 @@ require 'rails/generators'
 require 'rails/generators/migration'
 require 'rails/generators/active_record'
 
-
-module DataMigrations 
+module DataMigrations
   class InstallGenerator < Rails::Generators::Base
     include Rails::Generators::Migration
     source_root File.expand_path("../templates", __FILE__)
@@ -13,11 +12,11 @@ module DataMigrations
     def self.next_migration_number(path)
       Time.now.utc.strftime("%Y%m%d%H%M%S")
     end
-    
+
     def copy_initializer
       migration_template "create_data_migrations.rb",
                          "db/migrate/create_data_migrations.rb"
     end
-  end    
+  end
 end
